[autotest][cfm] Improve webview context polling
Previously we were just doing a get() of the webview context  which
worked for single screen setups, but for dual screen setups it
failed since the contexts are being destroyed and recreated by the
Hotrod app during OOBE if running with a MIMO (which is WAI).
To fix this we now poll repeatedly for the correct webview context
and also ignore the slave screen (the one NOT showing the OOBE screen)
to avoid exceptions (e.g. trying to interact with an already destroyed
context).

TEST=local with MIMO
BUG=chromium:772826

Change-Id: Ifa7fe584ff2a3565f4c2d5442ec34acb32013d8d
Reviewed-on: https://chromium-review.googlesource.com/721640
Commit-Ready: Denis Tosic <dtosic@chromium.org>
Tested-by: Denis Tosic <dtosic@chromium.org>
Reviewed-by: Kristoffer Erlandsson <kerl@google.com>