[autotest] Exclude hosts in more than one pool from lab inventory.
The lab inventory script fails if a DUT is assigned to a non-managed
pool.  The query that constructs the inventory selects only hosts
that are in managed pools, but failure may still occur if a DUT is
in both a managed and non-managed pool.

The lab inventory script isn't meant to allow DUTs to be in more
than one pool, so this excludes those DUTs from inventory.

BUG=chromium:757815
TEST=run the script locally

Change-Id: Ia5f5ed762b58e9bd6ce3014c07ee7cc8fd39887c
Reviewed-on: https://chromium-review.googlesource.com/724216
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>