Fix number of cpus return value for webrtc stats.
We can still not parse the new flashlight stats. It seems it is due to
the test expecting the number of CPUs to be a scalar value, not a list
as the other metrics.

Also added logging of exceptions that cause test failures for easier
troubleshooting.

BUG=chromium:774414
TESTED=Local DUT
pre-cq-configs: bluestreak-pre-cq

Change-Id: Ie01eb6028f81d14e208379bc09b3d915b2237811
Reviewed-on: https://chromium-review.googlesource.com/728740
Commit-Ready: Kristoffer Erlandsson <kerl@google.com>
Tested-by: Kristoffer Erlandsson <kerl@google.com>
Reviewed-by: Denis Tosic <dtosic@chromium.org>