Improve performance of gbm.crtcScreenshot()
I observed that the function was taking about 6000 ms to capture the
screen on an X86 Chromebook. The implementation was using a Python
bytecode function to copy and convert the bytes from the raw framebuffer
memory.

By comparison, this replacement takes about 45 ms. The speedup comes
from leveraging the built-in pixel format conversion code for the PIL
image library that was already being used (25x faster), and also by
making a simple copy of the raw framebuffer bytes using the Python
standard library (an additional 5x faster), and passing that buffer PIL.

By making this function so much faster, there is now a much smaller
window where the framebuffer might be written to while it simultaneously
being read by the autotest framework. The copied framebuffer image will
be much more likely to be consistent, and the cost of grabbing another
copy if unexpected pixels are observed will be much cheaper.

TEST=Observed a speedup and more consistent images in a new test
BUG=b:64448956

Change-Id: I6b95e2674f7b834719c51f682237464cd6dc5f10
Reviewed-on: https://chromium-review.googlesource.com/729325
Commit-Ready: Lloyd Pique <lpique@chromium.org>
Tested-by: Lloyd Pique <lpique@chromium.org>
Reviewed-by: Joe Kniss <djmk@google.com>