multimedia: Raise the exception from client during a RPC call
When calling a RPC, any exception from client is lost. The
exception is converted to a string value. We used to raise
a general exception with messages of the call name.

This change parses the exception string value, and then
raises a proper exception, like TestFail, such that in the
test report, we can group tests together by their failures.

BUG=chromium:774737
TEST=Ran a test using RPC and checked the exception raised.

Change-Id: Ic9eeb3881f25f708d1c51c0f3a696e54b8beb1c0
Reviewed-on: https://chromium-review.googlesource.com/733849
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>
Reviewed-by: Denis Tosic <dtosic@chromium.org>