multimedia: Suppress the close_tab RPC error
The close_tab RPC is usually called after a screen test is done, in
a finally clause. If there is an exception happened during the
screen test, the close_tab RPC may be failed too (as the tab may not
be openned). The close_tab exception (in the finally clause) will be
raised, instead of the screen test exception which we care.

We should ignore the close_tab exception in this case.

BUG=None
TEST=Ran a chameleon test with an exception raise in the screen test.

Change-Id: I696f7d7ca8780327fee40dee79674aea402804d9
Reviewed-on: https://chromium-review.googlesource.com/733811
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>