Fix issue in autotest power_cycle_utils
The power_cycle_gpio function will check whether /sys/class/gpio/gpioxxx
exits before toggling its value. There is the bug in the code that the
check is done on the host machine instead of the DUT. So when the file
does exit, the test will fail. The CL is meant to fix this.

BUG=chromium:776141
TEST=run the autotest locally and verify it passes with the gpioxxx file
exit

Change-Id: I29363f799ec119b4b13a7f37d755add4f0731a54
Reviewed-on: https://chromium-review.googlesource.com/734187
Commit-Ready: Zhongze Hu <frankhu@google.com>
Tested-by: Zhongze Hu <frankhu@google.com>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>