Open chrome://hitograms after a video is ended in ChromeVidResChangeHWDecode test
ChromeVidResChnageHWDecode opens chrome://histograms to check
Media.GpuVideoDecoderInitializeStatus value while a test video is played.
Because of that, the video played tab becomes background.
Since Chrome pauses a video in background tab, the test is failed because it
detects a test video is unexpectedly stopped.
To fix this failure, the test opens chrome://histograms after a test video is ended.

BUG=chromium:768621
TEST=video_ChromeVidResChangeHWDecode on caroline

Change-Id: Id62970a91d1ff416820812c1d45ddd34c1762746
Reviewed-on: https://chromium-review.googlesource.com/734881
Commit-Ready: Hirokazu Honda <hiroh@chromium.org>
Tested-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Kuang-che Wu <kcwu@chromium.org>