abstract_ssh.py: Fix wait_down() always waiting the full timeout
Adds a max_ping_timeout parameter to wait_down() that limits the
timeout on individual calls to get_boot_id(). The default is 10 seconds.

Before, each ping of get_boot_id() had a timeout of
  (full timeout) - (time already waited),
which means that it always waited for the full timeout in case the
host went down, no matter when the host actually went down.

As a result, this should significantly reduce the runtime of every
autotest that reboots the machine. For instance, the runtime of
enterprise_KioskEnrollmentServer, which calls ClearTPMOwnerRequest twice
triggering two reboots, reduces from 4m20s to 3m20s. It reduces even
further by another minute if the user presses CTRL+D to reboot the DUT
instead of waiting 30 seconds for automatic reboot.

BUG=None
TEST=test_that -b <dut_board> <dut_ip> enterprise_KioskEnrollmentServer

Change-Id: I7ead630c44348e9ffad96597ce971fcbb3b8c1f2
Reviewed-on: https://chromium-review.googlesource.com/741704
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Lutz Justen <ljusten@chromium.org>