gs_offloader: Add a counter for pass/fail on exit.
This counter is triggered on exit of the main loop, it will count
clean/dirty exits. Since we can't currently do a clean exit, there
we'll never record that, but the metric will be inplace if useful
later on. Counting dirty exits is useful now, since we want to alert
when the devserver starts crashing regularly.

BUG=chromium:773054
TEST=None

Change-Id: Ibf23a1d82b21b88df153174ab5cc379569698780
Reviewed-on: https://chromium-review.googlesource.com/742361
Commit-Ready: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>