tradefed_test: replace reboot with restart ui.
We used to reboot at the end of each tradefed iteration. This takes some
time (about 30s) and has a small risk of losing the DUT. Replace the
reboot in most cases with a browser restart. This should still give us a
fairly clean system.

Future optimization might include not restarting the browser at the end
of the test and reusing Android if found running at the beginning of the
test. But I will keep this for a future change.

BUG=b:62282232
TEST=time test_that <minnie> --iteration 10
     cheets_CTS_N.7.1_r10.arm.CtsDramTestCases
     (All pass.)

Change-Id: Id4970e08954e49cd0646bde02656c62549f4a32f
Reviewed-on: https://chromium-review.googlesource.com/742228
Commit-Ready: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>