Add firmware upgrade stress test for Atrus
Test to repeatedly upgrade the Atrus device, and verify that upgrade
work as intended. The upgrade is classified as successful if the
transfer of the firmware file, and the writing of the firmware file
to flash on the device was successfull.

BUG=b:68496712
TEST=Tested on a local DUT setup.
pre-cq-configs: bluestreak-pre-cq
CQ-DEPEND=CL:677389

Change-Id: Iffc5257cfb0acc0e478ceead601274f520653f1e
Reviewed-on: https://chromium-review.googlesource.com/744010
Commit-Ready: Marcus Wirebrand <wirebrand@chromium.org>
Tested-by: Marcus Wirebrand <wirebrand@chromium.org>
Reviewed-by: Kristoffer Erlandsson <kerl@google.com>