multimedia: Append the RPC name on the exception message from client
Append the RPC name on the exception message (TestFail or TestError)
from client. It is helpful with debugging. An example is like:
  FAIL: Failed to set mirrored (RPC: display.set_mirrored)

Also make other exceptions with similar messages, like:
  FAIL: Unhandled Exception: DevtoolsTargetCrashException: Devtools
  target crashed (RPC: cfm_main_screen.end_hangout_session)

If the original exception type is failed to be parsed, raise the
default one, like:
  FAIL: Unhandled Exception: Exception from client (RPC:
  cfm_main_screen.enroll_device)

BUG=chromium:774737
TEST=Ran a test using RPC and checked the exception raised.

Change-Id: Icb5d779d994702af65af646e6cb801321e439647
Reviewed-on: https://chromium-review.googlesource.com/744522
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>