bluetooth: check chameleon host
It is more user friendly to raise a test error in the beginning of a
test when the chameleon_host argument is not specified in a test
that requires the specification of the chameleon host. When the
argument is not specified, it would emit a message as follows
which is confusing.

    tko parser:   'MOUSE': lambda host:
                  host.chameleon.get_bluetooth_hid_mouse,
    tko parser:   AttributeError: 'NoneType' object has no attribute
                  'get_bluetooth_hid_mouse'

With this patch, it would emit a clearer message as follows if the
argument is omitted.

    [  FAILED  ]    ERROR: Have to specify chameleon_host IP.

BUG=chromium:780367
TEST=Conduct the following tests

(cr) $ test_that $DUT_IP bluetooth_AdapterPairing.mouse
This should raise an error such as
bluetooth_AdapterHIDReports.mouse  ERROR: Have to specify chameleon_host IP.

(cr) $ test_that --args "chameleon_host=$CHAMELEON_IP" $DUT_IP bluetooth_AdapterPairing.mouse
This should pass.

Change-Id: I5ed16f68b2bd5d3eb7ae71bdfe250f529d448f08
Reviewed-on: https://chromium-review.googlesource.com/752784
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Ruchi Jahagirdar <rjahagir@chromium.org>