bluetooth: skip rebooting chameleon
Sometimes, the chameleon host would automatically reboot when a
test failed multiple times with an attempt to fix issues caused
by unsuccessful detection of bluetooth peripherals. These kind
of issues are usually related with hardware.

However, such reboot consumes much test time and would bring
more troubles especially when it occurs in the middle of a
bluetooth pairing or connection. All states would be lost.
The consequence is that the following tests would definitely fail
due to the lost states.

Let's disable the reboot for the time being until we have a
better way to fix the lost states issue.

BUG=chromium:780367
TEST=Observe that the chameleon host does not reboot even when
tests failed multiple times.

Change-Id: If35a83ca5d1300ad099e1ad613b0ae0e28a89127
Reviewed-on: https://chromium-review.googlesource.com/752786
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Ruchi Jahagirdar <rjahagir@chromium.org>