[autotest][cfm] Adds server-side smoke test for Autotest framework
Used to test/track the stability of the Autotest framework when running
server side tests. The test starts a browsers, open 'chrome://version'
in a tab and then closes it. If the test fails it indicates a problem
in the core test infra.

BUG=chromium:771066
TEST=Tested against local setup.
pre-cq-configs: bluestreak-pre-cq

Change-Id: I46e7e1ee0db942665c10f0dbae86d73089f469fb
Reviewed-on: https://chromium-review.googlesource.com/749323
Commit-Ready: Denis Tosic <dtosic@chromium.org>
Tested-by: Denis Tosic <dtosic@chromium.org>
Reviewed-by: Kristoffer Erlandsson <kerl@google.com>