component_UpdateFlash: Check for clean shutdown.
This updates the component_UpdateFlash check to test for clean
shutdown. The bug causing the dangling reference counts on unmount only
occurs if the UI actually loads and uses the flash library, so this case
needs to be checked in the actual component flash test.

BUG=chromium:779273
TEST=emerge-${BOARD} autotest-tests

Change-Id: I6d2cafe19e673e4df179a04f609a0b80af51f27d
Reviewed-on: https://chromium-review.googlesource.com/756321
Commit-Ready: Greg Kerr <kerrnel@chromium.org>
Tested-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>