Remove __attribute__((overloadable))
The new FORTIFY removes this attribute from the non-FORTIFY'ed variants
of open. In order to keep clang happy, we need to remove it from here,
too.

Since system headers updates are apparently really difficult to do at
the same time as other changes, this includes a hack to make this file
compatible with old and new FORTIFYs. The hack (not the whole patch) is
intended to be reverted when the new FORTIFY has settled.

BUG=None
TEST=Builds with the new FORTIFY

Change-Id: I48e7211983bde200c37edcb0e5cc2f5f349cfd12
Reviewed-on: https://chromium-review.googlesource.com/756291
Commit-Ready: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Luis Lozano <llozano@chromium.org>