firmware_programmer: Use servo_host to execute commands
Introduce a servo_host as an argument (default to the servo_host
object from the servo object) and use it to execute commands.

It allows passing a CrosHost object to it and execute commands
on DUT, instead of Beaglebone.

BUG=b:69160863
TEST=Ran provision_FirmwareUpdate which executes commands on
servo_host.

Change-Id: I4fd885e045d8be7264e8bc6cc2aa27d5374496e8
Reviewed-on: https://chromium-review.googlesource.com/782716
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>