tpm_utils: wait for TPM ready after clearing the TPM owner
Add an option to wait for the TPM to show that it is ready after
clearing the TPM owner.

There are some tests that use cryptohome after clearing the TPM owner.
There was a change that prevented cryptohome from starting for a while
after the TPM owner was cleared. Those tests started failing because
cryptohome was not running right after reboot. For those tests not
verifying cryptohome, they can use wait_for_ready=True to make sure that
they wont fail because of issues like this.

BUG=b:69056272
BRANCH=none
TEST=run firmware_FWMPDisableCCD on R63-10032.27.0.

Change-Id: I0cbac86d0f3bd3bfdb265ab29956389ef5091c39
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/769427
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>