autotest: update SPINLOCKUP expectation for crash
SPINLOCK test provokes a crash and on reboot searches
for 'softlockup: hung tasks' in /dev/pstore.

However since the below patch,
9e300c7e3df3 CHROMIUM: scheduler: panic if schedulin while atomic
on provoking a crash a BUG() is introduced due to which
the crash is not recorded in a manner that SPINLOCKUP expects
i.e., 'softlockup: hung tasks'
instead its recorded as "BUG: scheduling while atomic"

But its not very consistent either.
Hence this patch adds both strings to be searched in crash log,
as a message as expectation for the test to pass.

Since 9e300c7e3df3 is _intended_ to reboot the system
on a warning, reverting it is not suggested for Chrome OS.

Hence this patch modifies the expectation of test.

BUG=b:69179392
TEST="test_that <ip> platform_KernelErrorPaths.SPINLOCKUP" passes

Change-Id: Id2e251aad8e29af94c9478fa232756a51977922d
Signed-off-by: Shirish S <shirish.s@amd.com>
Reviewed-on: https://chromium-review.googlesource.com/776637
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>