[autotest] Fix is_available()
So there are many different kinds of Host objects, and I put this
method on the wrong one.  Both of these are from AFE RPC calls and
thus use the RPC return Host object, not the Django database Host
object or the scheduler database Host object, or any of the two other
types of Hosts.

BUG=chromium:787233
TEST=None

Change-Id: I1ce07b2b827fde301707b14d85e658ce2f4c427f
Reviewed-on: https://chromium-review.googlesource.com/780844
Commit-Ready: Allen Li <ayatane@chromium.org>
Tested-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Ningning Xia <nxia@chromium.org>