chameleon: Unplug the monitor to skip video check during reboot
The use_edid_file() call for switching the EDID of the emulated monitor,
implicitly unplugs the monitor first, applies the EDID, replugs the monitor,
and then verifies the video signal.

For the reboot test, the video signal is not expected to be correct
during reboot. So skip the check by explicitly unplug the monitor before
reboot.

BUG=b:69528719
TEST=Ran the display_Resolution.reboot test on Peach_Pi

Change-Id: Ifb4bcc98e3cebb6705cad04d2bbaa574e1452c70
Reviewed-on: https://chromium-review.googlesource.com/798112
Commit-Ready: Wai-Hong Tam <waihong@google.com>
Tested-by: Shirish S <shirish.s@amd.com>
Tested-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>