tpm_utils: wait for attestation after clearing tpm owner
Wait for both the attestation and tpm to be ready after clearing the tpm
owner if wait_for_ready is requested. After the tpm is ready, prepare
for enrollment is run. This takes a long time, and occupies cryptohome
for a while. If you try to remove rootfs verification or use crypohome
during this, it will fail. Wait for attestation to be prepared before
returning, so we know prepare for enrollment is done and cryptohome can
be used.

BUG=none
BRANCH=none
TEST=run firmware_Cr50Update a bunch

Change-Id: I26671b70c98d39f0dcdc0302336f61a9fa3d1490
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/802398
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>