video_VideoCapability: add a test for verifying autotest-capablity settings for video features
We are going to use static capabilities in /etc/autotest-capability so that
each autotest decides if the test should be run on the DUT. If the device doesn't
satisfies a capability required by the test, the test is skipped and returns TEST_NA.
video_VideoCapability is a test for verifying autotest-capability settings.
Additionanly, it will be used to monitor capabilities on each DUT in the future.

BUG=chromium:789866
TEST=video_VideoCapability on stout

Change-Id: Ie6a0e6e3206d53c060a0a4b87c1b648c611e04b2
Reviewed-on: https://chromium-review.googlesource.com/814254
Commit-Ready: Hirokazu Honda <hiroh@chromium.org>
Tested-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Kuang-che Wu <kcwu@chromium.org>