Add remora enrollment autotest for display usage
The enrollment screen should be shown on a particular display (Mimo) in
the remora case. This is implemented in Chromium by making the Mimo the
primary display during enrollment.

This CL add a test case ensuring that the Mimo is the primary display
during remora enrollment.

BUG=chromium:795722
TEST=Ran the new test on a panther device.

Change-Id: Ibd21bb04d74945aed3576696fd353c2f9eab0a50
Reviewed-on: https://chromium-review.googlesource.com/847574
Commit-Ready: Felix Ekblom <felixe@chromium.org>
Tested-by: Felix Ekblom <felixe@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>