[autotest] Update client fail status line regex.
The regex was too greedy and absorbing any ':' that was past the
localtime info. Made the regex explicit about what was localtime so
that we don't lose everything after localtime that has a ':' in it.

BUG=chromium:782299
TEST=test_that dummy_FailServer, I modified dummy_FailServer to call
dummy_Fail as a client side test and saw that the test failure message
was properly pulled into dummy_FailServer failure message even if the
failure message was multiple lines long and had ':' in it.

Change-Id: I6619249afcd9aee7852a9def48028d4caefeca20
Reviewed-on: https://chromium-review.googlesource.com/804115
Commit-Ready: Kevin Cheng <kevcheng@chromium.org>
Tested-by: Kevin Cheng <kevcheng@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>