[autotest] Add a test for forced update at OOBE.
We want to automatically test the first update experience for many
users at OOBE. This CL adds a client test to start the update and a server
test to verify that update completes successfully.

The test gets a payload to use from the job_repo_url. This will give us
a payload for updating to the build that we are currently running
(N->N). This is because most of time this test runs it will be on a ToT
build.

Updates at OOBE are forced only when the 'deadline' option exists in the
omaha response. We will start a devserver with the --critical_update
flag to do this.

We create a custom lsb-release file to override the default. This will
tell update_engine to ping our custom devserver at OOBE.

The test will later be expanded to reboot in the middle, sleep, kill
network etc. So we could not have the devserver it communicates with be
on the DUT.

BUG=chromium:782828
TEST=test_that <ip> autoupdate_ForcedOOBEUpdate

Change-Id: Ifee5c9389320e8baf4a8998649de18abc182050d
Reviewed-on: https://chromium-review.googlesource.com/804602
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>