Add OmahaDevserver back to autotest libraries.
This Class was in use with autoupdate_EndToEndTest for years. It was
removed when that test was moved to the devserver-on-dut model. Regular
update tests can use the local devserver model.

However, for updating at OOBE we need a devserver instance that can
survive a DUT reboot because:
1. This update reboots in the middle without user intervention
2. we will want to extend this test to reboot in the middle of the update
and verify that it continues. (We never want users to be able to skip forced
update).

We cannot use exisitng lab devservers because we need devservers with two
extra flags: --critical_update and --host_log. The first ensures that
deadline:now is set in the omaha response which trigger forced update.
The second ensure we have update events to parse and verify. These two flags
are not used to start lab devservers.

BUG=chromium:782828
TEST=None

Change-Id: I3308d3b1c060d45262de4106fa8ded27461a603f
Reviewed-on: https://chromium-review.googlesource.com/805054
Commit-Ready: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>