autotest: add two new tables and one new column for afe_labels migration
Add two new tables: afe_static_labels and afe_static_hosts_labels. Add
one new column, replaced, to afe_labels, which means the value of this
label is replaced by the corresponding one in afe_static_labels. This CL
is one step to migrate the local afe static labels data to inventory
service.

BUG=chromium:791808
TEST=run the following command on a test ganeti server:
$AUTOTEST_DIR/database/migrate.py sync --debug
The new changes are in the database.

Change-Id: Ie82d764926d2372924588a000b3b5715f872bccf
Reviewed-on: https://chromium-review.googlesource.com/807528
Reviewed-by: Dan Shi <dshi@google.com>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Shuqian Zhao <shuqianz@chromium.org>