[autotest] Remove suite:perf_v2
This suite only has one test remaining in it (platform_MemoryMonitor).
That test creates very many timeseries which weren't particularly useful
on a performance dashboard, and, hence the entire suite was de-scheduled
from running in the lab.

In any case, the recommend way to track performance is now by tests
running in the suite:bvt-perbuild.

Therefore, let's completely kill the vestiges of this suite.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:791830
TEST=none

Change-Id: Ie073bcf3694111c7bad69d627ae6826485c2ee28
Reviewed-on: https://chromium-review.googlesource.com/807638
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>