[autotest] Remove platform_MemoryMonitor
This test is currently broken due to a change in the output of
"top -b -n 1".	 While we could fix this test to properly parse the output,
or preferably, modify it to fetch memory values (MemTotal, MemFree,
SwapTotal, etc.) from a more stable source (eg /proc/pid/stat), this test
is fundamentally broken anyway because of the way it generates a ton of
individual timeseries which cannot be tracked in a dashboard.

In fact, this problem led to its suite (suite:perf_v2) being de-scheduled
in 2016.  Since nobody has complained about the missing data for 2 years,
let's just completely remove the test.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:791829
TEST=None

Change-Id: I693845a118fa3029f042bc55f108d81a8cf65437
Reviewed-on: https://chromium-review.googlesource.com/807744
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>