Add a minimal server side test suite excluding SSP.
We do not support SSP in moblab-generic-vm yet. The minimal dummy_server
suite runs dummy_PassServer thrice -- with SSP, without SSP and maybe
with SSP. This is wasteful in the case when we don't support SSP at all.
So create an even simpler suite with just the nossp version of the test.

We still want to run it via a suite, so that we test the dynamic_suites
logic.

BUG=chromium:793016
TEST=moblab-vm-tests.

Change-Id: I803076f11091fd5920e422de80d83a95d3ca3879
Reviewed-on: https://chromium-review.googlesource.com/815134
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>