Use the ExportMediaInfo mod to for media metrics.
Instead of parsing JavaScript logs.

As a first step this is enabled by checking the is_meeting flag in
enterprise_CFM_Perf. The intention is to separate the meet test
to an own test as a next step.

the media_info_metrics_extractor and much of media_metrics_collector is
copied from Google3 and adopted to suite the Autotest style.

TEST=Local DUT
BUG=chromium:788703
pre-cq-configs: bluestreak-pre-cq
Change-Id: I98a80c0280a26074e98304b1510a6c198c6157d8
Reviewed-on: https://chromium-review.googlesource.com/817557
Commit-Ready: Kristoffer Erlandsson <kerl@google.com>
Tested-by: Kristoffer Erlandsson <kerl@google.com>
Reviewed-by: Claes Malmnäs <malmnas@google.com>
Reviewed-by: Denis Tosic <dtosic@chromium.org>