Remove some tests from suite:kernel_per-build_regression
We would like to start scheduling suite:kernel_per-build_regression in the
Chrome OS lab again to get some test coverage.  We are still in the
process of analyzing which tests really belong in the suite.

For now, remove the tests that have not yet been vetted or approved

BUG=b:69807272
TEST=test_that --pretend ${DUT} suite:kernel_per-build_regression
  => Does not include any of the tests removed by this CL

Change-Id: I013c8f1b62acca23976d7c6489660cf4c34df58f
Reviewed-on: https://chromium-review.googlesource.com/820562
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Commit-Queue: Daniel Kurtz <djkurtz@chromium.org>