container_pool: Remove client reliance on cxn
Client has a bare wait for a return from the server, relying on a
server-side guarantee that a return will be sent within the timeout.

This is not robust to problems with the connection, so this adds a
timeout on the client side. This changes it to a configurable
server-side timeout for a pool to be created, plus a client-side added
timeout for connection issues after which a return is presumed
impossible and will be ignored.

BUG=chromium:794026
TEST=Added to client_unittest

Change-Id: I829fc9aa988773962ee1648c287de696a3454976
Reviewed-on: https://chromium-review.googlesource.com/823643
Commit-Ready: Jacob Kopczynski <jkop@chromium.org>
Tested-by: Jacob Kopczynski <jkop@chromium.org>
Reviewed-by: Ben Kwa <kenobi@chromium.org>