faft/firmware_EventLog: Remove wait_for_client_offline() during recovery
The wait_for_client_offline is unecessary after doing a recovery boot.
The issue is that if the power_state:rec command in the line prior
takes a long time to return, then the DUT could be booted up already
by the time the wait_for_client_offline() command is issued, thus
causing the test to fail when in fact we do want a valid recovery boot
from the USB, and thus causing the test to fail when it shouldn't.
Note that the original test was booting into the recovery fw screen
whereas we're changing the test to now boot into recovery and boot up
the USB stick.  Both will get us the recovery boot entry in the
eventlog that we are checking for.

BUG=b:69567837, b:69860420
BRANCH=None
TEST=run firmware_EventLog on fizz with DUT booted into both
     normal and dev modes.

Change-Id: Id63d57d289f9d1e320a84703e6ec05c746f7d3f1
Signed-off-by: Shelley Chen <shchen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/825764
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Julius Werner <jwerner@chromium.org>