chrome_cr50: compare entire level string
Before we were using 'level not in current_level' to compare the
requested level to the current one. This didn't work as intended because
two valid privilege levels are lock and unlock. lock is unlock, so this
comparison doesn't work. This changes all of the level checks to be '=='
or '!='. This change also adds get_ccd_level to standardize ccd get
level until hdctools changes fixing servo cr50_ccd_level can land.

BUG=none
BRANCH=none
TEST=none

Change-Id: I7f2c99c293f833f928374b25439a9e2b47db1c24
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/828042
Reviewed-by: Wai-Hong Tam <waihong@google.com>