Use cros_config as source for test model label
With unified builds, there has been an ongoing mis-alignment between how
models are configured in the build and how we actually want to test
them.  Added new test-label attribute to cros_config, which will be used
if present in Autotest to set the model label.  This will allow
platforms where we don't care about the difference to be treated as and
reported as the same (e.g. electro/basking, robo/robo360).

BUG=chromium:789475
TEST=cros_label_unittest

Change-Id: I5ab0b3e03471db1868045602053d17c29bedd129
Reviewed-on: https://chromium-review.googlesource.com/828150
Commit-Ready: C Shapiro <shapiroc@google.com>
Tested-by: C Shapiro <shapiroc@google.com>
Reviewed-by: Jason Clinton <jclinton@chromium.org>