cr50_stress_flaky: remove firmware_Cr50Update.erase_nvmem
nvmem is reinitialized on the transition from dev to prod.
firmware_Cr50Update updates to the dev image and then rollsback to the
prod image to do switching to older images. Because this dev to prod
transition is built into the test, we don't need
firmware_Cr50Update.erase_nvmem.

I am leaving the support for erasing nvmem, because sometimes we verify
the release before signing it with the prod key, and it could be useful
to erase nvmem on the rollback.

BUG=none
BRANCH=none
TEST=none

Change-Id: I007a5eb686a18d996069bc100f086bb4647d06c9
Signed-off-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/832986
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>