tradefed_test: tune login timeouts.
We are interested in fairly tight login timeouts for the CQ to not
wait too long in case of problems. But it appears that we need to
be able to relax the login timeout for some boards like betty,
which can run on slow GCE instances.

This change
- increases the regular Chrome login timeout from 60s to 90s.
- increases betty timeout from 60s to 300s.

BUG=chromium:794707
TEST=pre-cq will test.

Change-Id: Ifeea56cd609395a052ea7ee059de450a504b73b2
Reviewed-on: https://chromium-review.googlesource.com/833502
Commit-Ready: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>