tradefed_test: Retry decision based on 'list result'.
Previously tradefed_test performed retries based on the result parsed from
the tradefed 'run' command output. The approach was fragile; for instance
it was inherently impossible to count "not executed" tests, hence it
sometimes marked a failing run as passing (b/65131971). Or, it could not
retry further when a retry run aborted before running any test because
because it loses track of the session id in this case (b/69719234).

With this CL, tradefed_test instead relies on the tradefed 'list results'
command output. It gives us precisely the info that tradefed uses to decide
the need of retries.

Drawback of the new approach is that we lose the double integrity-check.
When tradefed did something wrong, previously tradefed_test detected
the test count anormaly, but now it won't. I hope this is not a big
issue. Historically, all the recent alerts from the check were false
alerts (b/69721080), and now we also have the test count info in the CTS
dashboard by using 'collect-tests-only' as well.

BUG=b:69719234
BUG=b:69721080
BUG=b:65131971
TEST=Manually run many Cts/Gts tests w/ artificially introduced failures.

Change-Id: I278d4edfe6312efbd2c0ecf4fca67403537f7c2e