[autotest] Track COMPLETED event for lucifer
If lucifer_run_job crashes, we previously could not detect it.
Non-zero exit status may not mean a crash; it could mean that the test
failed or some other recoverable error.

lucifer_run_job is Go, so the most common crash would result in an
exit status above 127 (due to a signal, e.g. SIGSEGV).  However, that
is non-portable, and there are other ways for Go to crash (e.g. panic,
log.Fatalf()).  Thus, relying on COMPLETED seems best.

lucifer_run_job is responsible for driving the job to completion,
whatever that entails, and emitting COMPLETED.  If it cannot, then
\_()_/ and let job_aborter clean up the mess.

BUG=chromium:794779
TEST=None

Change-Id: I32ffaf5b2906ccec4b4086f1081f28d286aafda0
Reviewed-on: https://chromium-review.googlesource.com/837976
Commit-Ready: Allen Li <ayatane@chromium.org>
Tested-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>