moblab_RunSuite: Add test argument to hint at overall timeout.
For this test, cleanup action is very important. In case the test is
about to timeout, it is important to abort the suite within moblab and
collect logs from the moblab. A new test argument lets us hint at the
overall test timeout.

This will be particularly useful for MoablabVMTest where the test is run
via test_that.

BUG=chromium:799199
TEST=moblab_RunSuite on moblab-vm, with and without the new argument.

Change-Id: I58bfda46c5e0fe648e04b77fbb00bc57321fb69b
Reviewed-on: https://chromium-review.googlesource.com/851119
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>