autotest: make BootLockbox test a server-side test
The client-side platform_BootLockbox to run requires that the
boot lockbox on the device is not locked yet. Any user sign in locks
it. So, if the test is run on a device where a user sign in has
already been performed (e.g. as a part of other test), it detects
that the boot lockbox is locked and aborts (fails with a custom message,
added in CL:648570).

A simple way to ensure that the test can run is to reboot the device
before the test. This change creates a server-side
platform_BootLockboxServer test that reboots the device and then calls
the client-side platform_BootLockbox.

The change also replaces platform_BootLockbox in bvt-perbuild suite
with the new reliable platform_BootLockboxServer.

BUG=chromium:728714
TEST=login on DUT, run platform_BootLockboxServer, watch it reboot the
     device and pass.

Change-Id: Ia90478c4bbc9c626bd8d81352076282fdffafe9e
Reviewed-on: https://chromium-review.googlesource.com/852934
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>