moblab: Do not wait for AFE after reboot.
Not all callers of moblab_host.reboot want to wait for the AFE to be
back. In particular, moblab is rebooted twice during provision, and it
is irrelevant whether AFE comes back or not after these reboots.

In case of moblab-vm, the default timeout of 5 minutes isn't enough for
all moblab services to start (and hence AFE to come up), so this check
always fails. This adds 10 minutes of delay to provision,

Any users of moblab_host that reboot the device and are about to use the
AFE are aware that they're working with moblab. They should call
wait_afe_up explicitly. For example, MoblabTest does this already.

BUG=chromium:799724
TEST=moblab-vm-generic-pre-cq

Change-Id: I7cc448e00a8fb6a4093eba988a73edbee8250b27
Reviewed-on: https://chromium-review.googlesource.com/855199
Commit-Ready: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Keith Haddow <haddowk@chromium.org>