Fix emerge failure of autotest-chrome
This CL modifies the importing logic of mkvparse to make it dynamically
imported so that it won't be imported during test set up. This would
fix the problem that mkvparse library can't be found in host environment
during emerging the test.

BUG=616659
TEST=video_MediaRecorderPerf can be emerged and passes on kevin

Change-Id: I635319da91bf5262e28e0897fd1ce037fded1903
Reviewed-on: https://chromium-review.googlesource.com/855977
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Sheng-hao Tsao <shenghao@google.com>
Reviewed-by: Ricky Liang <jcliang@chromium.org>